requirements.txt
langchain
langchain-google-genai
playwright
apscheduler
praw
httpx
beautifulsoup4
python-dotenv
pyyaml
google-generativeai
supabase